BaseFragment moved into presentation package
