+++ app/src/main/java/com/mobven/fitai/presentation/home/HomeFragment.kt
@@ -6,7 +6,7 @@ import androidx.annotation.RequiresApi
 import androidx.navigation.NavDirections
 import androidx.navigation.fragment.findNavController
 import com.mobven.fitai.R
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.CardPlanItemBinding
 import com.mobven.fitai.databinding.FragmentHomeBinding
 import com.mobven.fitai.presentation.home.calendar.CalendarItem
