+++ app/src/main/java/com/mobven/fitai/presentation/home/screens/DietTypeFragment.kt
@@ -2,7 +2,7 @@ package com.mobven.fitai.presentation.home.screens
 
 import androidx.viewpager2.widget.ViewPager2
 import com.mobven.fitai.R
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentDietTypeBinding
 import com.mobven.fitai.presentation.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.presentation.login.sign_up.model.SignUpSelectorItem
