+++ app/src/main/java/com/mobven/fitai/presentation/home/screens/PlanCreatingFragment.kt
@@ -2,7 +2,7 @@ package com.mobven.fitai.presentation.home.screens
 
 import android.os.Handler
 import android.os.Looper
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentPlanCreatingBinding
 
 class PlanCreatingFragment :
