+++ app/src/main/java/com/mobven/fitai/presentation/login/sign_up/screens/AverageWaterFragment.kt
@@ -2,7 +2,7 @@ package com.mobven.fitai.presentation.login.sign_up.screens
 
 import androidx.viewpager2.widget.ViewPager2
 import com.mobven.fitai.R
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentAverageWaterBinding
 import com.mobven.fitai.presentation.login.sign_up.adapter.SignUpListAdapter
 import com.mobven.fitai.presentation.login.sign_up.model.SignUpSelectorItem
