+++ app/src/main/java/com/mobven/fitai/presentation/onboarding/OnboardingFragment.kt
@@ -1,6 +1,6 @@
 package com.mobven.fitai.presentation.onboarding
 
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentOnBoardingBinding
 
 class OnboardingFragment : BaseFragment<FragmentOnBoardingBinding>(FragmentOnBoardingBinding::inflate) {
