+++ app/src/main/java/com/mobven/fitai/presentation/onboarding/screens/FourthOnboardingFragment.kt
@@ -1,6 +1,6 @@
 package com.mobven.fitai.presentation.onboarding.screens
 
-import com.mobven.fitai.base.BaseFragment
+import com.mobven.fitai.presentation.base.BaseFragment
 import com.mobven.fitai.databinding.FragmentFourthOnBoardingBinding
 
 class FourthOnboardingFragment : BaseFragment<FragmentFourthOnBoardingBinding>(FragmentFourthOnBoardingBinding::inflate)
